Rename get API key credit v2 core module
