+++ src/cloud/get-api-key-credit-v2/handler.js
@@ -1,1 +1,4 @@
-export { createGetApiKeyCreditV2Handler, extractUuid } from './core.js';
+export {
+  createGetApiKeyCreditV2Handler,
+  extractUuid,
+} from './get-api-key-credit-v2-core.js';
